+++ Models/OrderDetail.cs
@@ -15,7 +15,7 @@ namespace TestWebbshopCodeFirst.Models
         public int Id { get; set; }
         public int OrderId { get; set; }
         public int ProductId { get; set; }
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int Quantity { get; set; }
 
         public virtual Order Order { get; set; } = null!;
